+++ snakeoil/process/__init__.py
@@ -12,97 +12,9 @@ from snakeoil.demandload import demandload
 demandload(
     'errno',
     'signal',
-    'subprocess',
     'time',
     'snakeoil.osutils:access',
-    'snakeoil.fileutils:readlines_ascii',
 )
-
-def _parse_cpuinfo():
-    data = readlines_ascii("/proc/cpuinfo", True, True, False)
-    procs = []
-    current = []
-    for line in data:
-        if not line:
-            if current:
-                procs.append(current)
-                current = []
-        else:
-            current.append(line.split(":", 1))
-    return [{k.strip(): v.strip() for k, v in items} for items in procs]
-
-def _get_linux_physical_proc_count():
-    procs = _parse_cpuinfo()
-    if not procs:
-        return _get_linux_proc_count()
-    core_count = {}
-    for proc in procs:
-        physical_id = proc.get('physical id')
-        if physical_id is None:
-            return _get_linux_proc_count()
-        if physical_id in core_count:
-            continue
-        cores = proc.get('cpu cores')
-        if cores is None:
-            return _get_linux_proc_count()
-        core_count[physical_id] = int(cores)
-
-    return sum(core_count.itervalues())
-
-def _get_linux_proc_count():
-    try:
-        return os.sysconf('SC_NPROCESSORS_ONLN')
-    except (ValueError, OSError, AttributeError):
-        procs = _parse_cpuinfo()
-        if not procs:
-            return None
-        return len(procs)
-
-def _get_bsd_proc_count():
-    p = subprocess.Popen(
-        ["sysctl", "-n", "hw.cpu"],
-        env={"PATH": "/sbin:/bin:/usr/sbin:/usr/bin"}, close_fds=True, shell=False,
-        stdout=subprocess.PIPE, stdin=None, stderr=subprocess.STDOUT)
-    p.communicate()
-    if p.returncode == 0:
-        try:
-            return int(out.strip() for out in p.stdin.read())
-        except ValueError:
-            pass
-    return None
-
-def get_physical_proc_count(force=False):
-    """return the number of non-HT cpu's identified
-
-    :param force: force recalculating the value, else use the cached value
-    :return: integer of the number of processors.  If it can't be discerned, 1 is returned
-    """
-    val = getattr(get_physical_proc_count, 'cached_result', None)
-    if val is None or force:
-        if 'linux' in sys.platform:
-            val = _get_linux_physical_proc_count()
-        else:
-            val = get_proc_count()
-        get_physical_proc_count.cached_value = val
-    return val
-
-def get_proc_count(force=False):
-    """return the number of cpu's identified, HT or otherwise
-
-    :param force: force recalculating the value, else use the cached value
-    :return: integer of the number of processors.  If it can't be discerned, 1 is returned
-    """
-    val = getattr(get_proc_count, 'cached_result', None)
-    if val is None or force:
-        if 'linux' in sys.platform:
-            val = _get_linux_proc_count()
-        elif 'bsd' in sys.platform or 'darwin' in sys.platform:
-            val = _get_bsd_proc_count()
-        if not val:
-            val = 1
-        get_proc_count.cached_result = val
-    return val
-
 
 def find_binary(binary, paths=None):
     """look through the PATH environment, finding the binary to execute"""